Restructured api to use services/handlers and starting replacing jwt with session auth
Moved functions from services into their own files. Also began replacing jwt auth with session auth.
