+++ api/internal/handlers/authhandler/getCurrentUser.go
@@ -2,7 +2,9 @@ package authhandler
 
 import (
 	"myanimevault/internal/models/customErrors"
-	"myanimevault/internal/services"
+	"myanimevault/internal/models/dtos"
+	"myanimevault/internal/services/useranimeservice"
+	"myanimevault/internal/services/userservice"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -10,9 +12,8 @@ import (
 
 func GetCurrentUserHandler(context *gin.Context) {
 	userId := context.GetString("userId")
-	userEmail := context.GetString("userEmail")
 
-	userDto, err := services.GetUserById(userId)
+	user, err := userservice.Get(userId)
 
 	if err != nil {
 		switch err {
@@ -25,23 +26,17 @@ func GetCurrentUserHandler(context *gin.Context) {
 		}
 	}
 
-	animeIdList, err := services.GetIdList(userId)
+	animeIdList, err := useranimeservice.GetIdList(userId)
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message": "There was a problem getting the users anime id list."})
 		return
 	}
 
-	token, err := services.GenerateAuthToken(userId, userDto.Email)
+	userDto := dtos.UserDto{}
 
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"message": "There was a problem generating an auth token."})
-		return
-	}
-
-	userDto.Id = userId
-	userDto.Email = userEmail
-	userDto.AuthToken = token
+	userDto.Id = user.Id.String()
+	userDto.Email = user.Email
 	userDto.AnimeIds = animeIdList
 
 	context.JSON(http.StatusOK, gin.H{"message": "Current user was successfully returned.", "user": userDto})
