+++ api/internal/models/dtos/userDto.go
@@ -3,6 +3,5 @@ package dtos
 type UserDto struct {
 	Id        string  `json:"id"`
 	Email     string  `json:"email"`
-	AuthToken string  `json:"authToken"`
 	AnimeIds  []int64 `json:"animeIds"`
 }
