+++ api/internal/services/sessionservice/create.go
@@ -1,0 +1,5 @@
+package sessionservice
+
+func Create(sessionId string) {
+	
+}
